All simple and 'multi-select' filters implemented. TODO Map based ones
Some initial integration with the UI, however the first filter click
event does not currently trigger. On inspection the issueVisibilities
map in the column component seems to be populated correctly, but the
boolean value passed in to the issue component does not get updated.
This is just a theory though :)