Implement the rank view and switching between the two views. Also
revisit backlog toggling.

If a view switch or backlog toggle means
- displaying the backlog and it is not currently shown
- hiding the backlog and it is currently hidden
we load up the full board again from the server before updating the
view. This is mainly to keep the issue rankings in the board state up to
date, since when changes come in they will be indexed according to
whatever mode we are in.