[#12] Server side for smarter linked issues
linked-projects now has a type-links map allowing overridden state
mappings by the type of the linked issues (for when the linked project
has different workflows for different issue types).

Inclusion of linked issues must now be explicitly turned on in the
main board projects (or in its overrides, which are by the issue type of
the issue referencing a linked issue). There is a filter to do filtering
by the link type, or the issue type, priority or labels of the linked
issue. For overrides the filter can enforce that no linked issues should
be included.

Add a type field to the linked issues if (and only if) it is a type
which has an override in the type-states of the linked-project.

Some cleanup