[#36] When the backlog is hidden don't add the backlog issues to the
issue table. This does not happen when running in Jira, but it does
happen when developing the UI and running the demo since for these they
get the same data both when showing the backlog and not (there is no
proper server).

Fix header tests to use a visible backlog so that the counts are the
same