Don't set UserSettingState.currentVisibleColumns
if they are the same as the default.

Toggle backlog even in rank view mode,

Together thse solve:

1) Problem where if in Kanban view show and then hide the
backlog, you end up with the backlog states hidden in the Rank view.

2) If you show the backlog in the Kanban view and go to the rank
view, and there collapse the backlog header/hide all backlog states,
when you come back to the kanban view the backlog is correctly hidden.
To show the backlog again you need TWO clicks, which seems to indicate
that the backlog header is in a strange state